Add files via upload
The first code snippet for this Jupyter Notebook is the feature processing for this code. The second is loading our data with a timeline file that automatically creates a graph  showing the alignment of the gesture activity points with our experimental protocol as described in the paper and its alignment compared to other sessions's alignment for this subject. Then is a list with 3 numbers, representing the approximate time point where your recording protocol started. For Subject 1's data the approximate start points we've used for sessions 1, 2, 3, were 2.0(s), 3.5(s), 1.75(s) respectively.  For Subject 2's data the approximate start points we've used for sessions 1, 2, 3, were 8.55(s), 8.3(s), 3.4(s) respectively. Towards the bottom of the code is the 2 MLP code snippets, the first code snippet will be the 6-Fold Stratified CV that splits the dataset randomly into 80% training and 20% for all sessions listed. This means that if you were to load 3 sessions from both subjects and give their respective protocol start times you can run this code on that dataset and it would split the data randomly into the 80% and 20% datasets, or instead if were to load 3 sessions from one subject it will take 80% randomly and test on 20% as well. Now for the second code, this is the cross session MLP evaluation using 5 k fold stratification which trains on the first 2 sessions you loaded and tests on the 3rd session. For this code there is also the option to save the predicted results to a CSV file. This is important for giving these predicted results to our prosthetic deployment code to replay the predicted session.